pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plaid2beancount"
version = "0.1.0"
description = "A tool to sync Plaid transactions to Beancount format"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Mark Reitblatt", email = "mark@reitblatt.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",    
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "plaid-python>=12.0.0",
    "beancount>=2.3.0",
    "flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools.dynamic]
dependencies = {file = ["requirements.txt"]}
[project.scripts]
plaid2beancount = "main:main"

[tool.setuptools]
py-modules = ["main", "plaid_models", "plaid_link_server", "transaction_models"]
packages = ["transactions"] 
